+++ lib/src/widgets/quill/text_line.dart
@@ -276,6 +276,9 @@ class _TextLineState extends State<TextLine> {
       Attribute.h1: defaultStyles.h1!.style,
       Attribute.h2: defaultStyles.h2!.style,
       Attribute.h3: defaultStyles.h3!.style,
+      Attribute.h4: defaultStyles.h4!.style,
+      Attribute.h5: defaultStyles.h5!.style,
+      Attribute.h6: defaultStyles.h6!.style,
     };
 
     textStyle = textStyle.merge(m[header] ?? defaultStyles.paragraph!.style);