+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 9.0.2-dev.2
+* Add the new translations for ru, uk arb files by [#1575](https://github.com/singerdmx/flutter-quill/pull/1575)
+* Add a new dropdown button by [#1575](https://github.com/singerdmx/flutter-quill/pull/1575)
+* Update the default style values by [#1575](https://github.com/singerdmx/flutter-quill/pull/1575)
+ 
+
 ## 9.0.2-dev.1
 * Add configurations for the new dropdown `QuillToolbarSelectHeaderStyleButton`, you can use the orignal one or this
 * Fix the [issue](https://github.com/singerdmx/flutter-quill/issues/1119) when enter is pressed, all font settings is lost